app/progress_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
PyJWT
requests
aiokafka==0.8.1
